CharScreen Setup
- Primitive Hub Character Screen Setup
- Texture Updates
- UI Prefab Updates

---
In this character screen, the player enters the menu and selects their character, then leaves. Since this operation only exists in the hub, there's no need for an explicit confirmation button since selecting the character is the action that actually changes it. We can use camera pans to support the idea.